+++ components/steps/index.ts
@@ -1,5 +1,5 @@
 import Steps from './Steps.vue'
-export type { Props, Step } from './Steps.vue'
+export type { Props, Item } from './Steps.vue'
 import { withInstall } from '../utils/type'
 
 export default withInstall(Steps)